Rearrange DBG_REG debug logging to reflect driver programatic access to the device registers.  Created a DBG_RGC to display internal changes to the registers.
Changed the behavior of device wen register changes happen to delay 20 instructions before acting on changed register state.  This is useful/necessary since some driver code writes some registers with the same value twice to be sure that the device notices the write.  We must male sure we don't try to act on the same write events more than once.