KMC: New baselevel
Not quite ready for production, but useful.

This baselevel is another major update.  TX and RX are now threaded.
Debug and trace have been reworked.
Many, many internal changes.

This version runs (with a hacked DUP) TOPS10, TOPS20 & RSX.

The DUP issues need to be resolved.

There are unexplained performance divots.

This version also contains an exerciser for the prototype enhanced help system.