PDF: Add timed flush support
The lp20 is the only current printer that flushes buffers/checkpoints .pdf files when
the printer is idle.  (Defined as at column zero for more than 10 seconds.)

This behavior is desirable for all printers, so centralize it in sim_pdflpt.

This means that sim_pdflpt will inspect data even if it is destined for a non-pdf file.

Remove the idle logic from pdp10_lp20 and let sim_pdflpt handle the function.

Fix oversight in (so-far unused) pdflpt_write.

Convert attach parameter parser to table-driven form and improve error detection.