PDF: Add support for rolling files.  Support image overlay
Device emulators want to switch to a new file, keeping the parameters of the old.
Added pdf_newfile to support this.  Expose pdf_open_exclusive for the wrapper
to use for non-pdf spooled files.

Images can now be specified with a form.  The image will be rendered over the form.
This allows logos to be applied to the standard forms.
For just the image, specify the Plain form.

Added capability to sim_pdflpt to manage non-pdf files when -S directs output to a spool
directory.  Combined with a watcher (not part of SimH, but a Windows demonstration is
provided), this allows 1-keystroke printing to a physical device.  See doc for details.

This also:
o Updates structured help to have a flat (outline) display option.  This can be controlled by the
   environment (e.g. non-interactive) and by a flag in the DEVICE (for those who prefer that format.)
   Even flattened, the outline provides some semblance of structure/organization.

o Adds '*' as an option to structured help.  This displays all subtopics in flattened form.
   HELP LP20 * will display all available help for the LP20 (currently 481 lines).

o Reformats magic topics to conform with structured help conventions.

o Corrects various bugs in SCP's help handling: it wasn't passing the command table flags to device
   help, logged to stdout twice (instead of once to sim_log), didn't allow device help to parse arguments,..

o Attempts to compress image data

o Provides a lpt2pdf  API to determine if a file has any content.

o Provides default help for lpt2pdf-controlled printers

o Updates LP11 and LP20 help

o Provides a simple spool file watcher for windows - not supported, proof of concept.

o Fixes a SimH crash caused by the new debug prompt.