LP20: Timing-accurate printers, auto forms length
Slow the LP20 down to device speeds.  Adds a baker's dozen real printers, and
a fake one for faster printing.

SHOW LP20 MODELS for a list
SET LP20 PRINTER-MODEL=name to select

Selecting a printer model selects speeds, VFU type (optical or DAVFU).

SET LP20 VFUTYPE replaced by SET LP20 PRINTER-MODEL and SET LP20 TAPE=file.

A VFU (of either type) defines the length of a form.  It now automagically  provides the length (defined as lines * lpi) to the PDF library as VFUs are loaded.  It also sets the tof-offset based on the BOF marker.  BOF to form length is the offset.

The OS default "NORMAL" VFU is 63 lines and expects a tof-offset of 3.  Automatically detect this

Added some DEBUG flags.

Removed STOPIOE - the OS can handle errors, and stopping the simulation is rude.