cmake: cmake build system
CMake-based build system that parallels the makefile-based supported
build system. It does not supercede, replace or otherwise eliminate
the makefile-based build system. It is merely an alternative means to
build simh with a slightly different approach to package dependencies on
Windows (and other platforms, as necessary).  All cmake configuration
information for all simulators is dynamically derived from the base simh
makefile which is the single point containing that information.

Documentation in README-cmake.md (work-in-progress)

This is derived from original efforts done by B. Scott Michel.  See
CMakeNotes.txt for comments.